refactor: further restructuring
- Renamed `about-us` components to `about`
- Moved `opportunities` to `pages/opportunities`
- `research-projects` components renamed to `projects`
- Updated `angular.json` and `app.routes.ts` to reflect changes
- Modified `footer.component.html` and `header.component.html` to reflect changes
