+++ angular.json
@@ -35,7 +35,7 @@
               "src/styles.scss"
             ],
             "scripts": [
-              "src/app/opportunities/Img.js"
+              "src/app/pages/opportunities/Img.js"
             ]
           },
           "configurations": {
@@ -95,7 +95,7 @@
               "src/styles.scss"
             ],
             "scripts": [
-              "src/app/opportunities/Img.js"
+              "src/app/pages/opportunities/Img.js"
             ]
           }
         }
