+++ src/app/pages/about/about.component.spec.ts
@@ -2,7 +2,7 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AboutUsComponent } from './about-us.component';
+import { AboutUsComponent } from './about.component';
 
 describe('AboutUsComponent', () => {
   let component: AboutUsComponent;
