+++ src/app/pages/projects/projects.component.scss
