+++ src/app/pages/projects/projects.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-projects',
+    imports: [],
+    templateUrl: './projects.component.html',
+    styleUrl: './projects.component.scss'
+})
+export class ProjectsComponent {}
