--- src/app/pages/research-projects/research-projects.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app-research-projects',
-    imports: [],
-    templateUrl: './research-projects.component.html',
-    styleUrl: './research-projects.component.scss'
-})
-export class ResearchProjectsComponent {}
