+++ src/app/shared/footer/footer.component.html
@@ -27,10 +27,10 @@
                             <div class="col-md-6">
                                 <ul>
                                     <li><a (click)="navigateAndReload('/home')">Home</a></li>
-                                    <li><a (click)="navigateAndReload('/research')">Research & Projects</a></li>
+                                    <li><a (click)="navigateAndReload('/projects')">Projects</a></li>
                                     <li><a (click)="navigateAndReload('/opportunities')">Opportunities</a></li>
                                     <li><a (click)="navigateAndReload('/gallery')">Gallery</a></li>
-                                    <li><a (click)="navigateAndReload('/aboutus')">About Us</a></li>
+                                    <li><a (click)="navigateAndReload('/about')">About</a></li>
                                 </ul>
                             </div>
                             <div class="col-md-6">
