+++ src/app/shared/header/header.component.html
@@ -29,10 +29,10 @@
       <nav class="main-nav" [class.menu-open]="isMenuOpen" [class.menu-closing]="isMenuClosing">
         <ul>
           <li><a (click)="navigateAndReload('/home')">Home</a></li>
-          <li><a (click)="navigateAndReload('/research')">Research & Projects</a></li>
+          <li><a (click)="navigateAndReload('/projects')">Projects</a></li>
           <li><a (click)="navigateAndReload('/opportunities')">Opportunities</a></li>
           <li><a (click)="navigateAndReload('/gallery')">Gallery</a></li>
-          <li><a (click)="navigateAndReload('/aboutus')">About Us</a></li>
+          <li><a (click)="navigateAndReload('/about')">About</a></li>
         </ul>
       </nav>
     </div>
